Add Contour Next One driver, refactor code for more contour devices
